refactor, comments and more
